requirements.txt
flask>=2.3.0
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
openai>=1.0.0
pyyaml>=6.0.0
pyinstaller>=5.0.0
